♻️ Refactor: ChannelGrid 네이밍 수정
ChannelGrid.tsx -> index.tsx
